Remove createTransferToUserBankTransaction and deriveUserBankAddress
